+++ cf_xarray/tests/test_accessor.py
@@ -479,7 +479,9 @@ def test_Z_vs_vertical_ROMS():
     from .datasets import romsds
 
     assert_identical(romsds.s_rho.reset_coords(drop=True), romsds.temp.cf["Z"])
-    assert_identical(romsds.z_rho.reset_coords(drop=True), romsds.temp.cf["vertical"])
+    assert_identical(
+        romsds.z_rho_dummy.reset_coords(drop=True), romsds.temp.cf["vertical"]
+    )
 
     romsds = romsds.copy(deep=True)
 
@@ -494,4 +496,41 @@ def test_Z_vs_vertical_ROMS():
     # use .coords if coordinates attribute is not available
     romsds.temp.encoding.clear()
     assert_identical(romsds.s_rho.reset_coords(drop=True), romsds.temp.cf["Z"])
-    assert_identical(romsds.z_rho.reset_coords(drop=True), romsds.temp.cf["vertical"])
+    assert_identical(
+        romsds.z_rho_dummy.reset_coords(drop=True), romsds.temp.cf["vertical"]
+    )
+
+
+def test_param_vcoord_ocean_s_coord():
+    from .datasets import romsds
+
+    romsds.s_rho.attrs["standard_name"] = "ocean_s_coordinate_g2"
+    Zo_rho = (romsds.hc * romsds.s_rho + romsds.Cs_r * romsds.h) / (
+        romsds.hc + romsds.h
+    )
+    expected = romsds.zeta + (romsds.zeta + romsds.h) * Zo_rho
+    romsds.cf.decode_vertical_coords()
+    assert_allclose(
+        romsds.z_rho.reset_coords(drop=True), expected.reset_coords(drop=True)
+    )
+
+    romsds.s_rho.attrs["standard_name"] = "ocean_s_coordinate_g1"
+    Zo_rho = romsds.hc * (romsds.s_rho - romsds.Cs_r) + romsds.Cs_r * romsds.h
+    expected = Zo_rho + romsds.zeta * (1 + Zo_rho / romsds.h)
+    romsds.cf.decode_vertical_coords()
+    assert_allclose(
+        romsds.z_rho.reset_coords(drop=True), expected.reset_coords(drop=True)
+    )
+
+    romsds.cf.decode_vertical_coords(prefix="ZZZ")
+    assert "ZZZ_rho" in romsds.coords
+
+    copy = romsds.copy(deep=True)
+    del copy["zeta"]
+    with pytest.raises(KeyError):
+        copy.cf.decode_vertical_coords()
+
+    copy = romsds.copy(deep=True)
+    copy.s_rho.attrs["formula_terms"] = "s: s_rho C: Cs_r depth: h depth_c: hc"
+    with pytest.raises(KeyError):
+        copy.cf.decode_vertical_coords()